refactored and added pytests
