+++ backend/player_management/tests/test_misc.py
@@ -8,6 +8,8 @@ from faker.providers import person, date_time
 from .models import Person, Club, PersonToClubMembership
 from django.test import TestCase
 from django_seed import Seed
+from mixer.backend.django import mixer
+import pytest
 import csv
 
 # Define Seeding function to quickly populate database for tests
@@ -40,34 +42,7 @@ def seed_person():
     fakePerson.firstname = fake.first_name()
 
 
+
 def loadcsvData():
     # TODO: Initialize data from the csv files. us the constructors for that
     pass
-
-class testDataModel(TestCase):
-    def setUp(self):
-
-        seeder = Seed.seeder()
-
-        seeder.add_entity(Person, 1)
-        seeder.add_entity(Club, 1)
-
-        inserted_pks = seeder.execute()
-
-        person_id = inserted_pks[Person][0]
-        club_id = inserted_pks[Club][0]
-
-        PersonToClubMembership(person_id=person_id, club_id=club_id)
-
-
-    def test_PersonToClubMembership(self):
-        """create a person and a club add person to club test isActive function"""
-
-        person = Person.objects.all()[0]
-        club = person.club_memberships.all()[0]
-        role = PersonToClubMembership.objects.get(person=person, club=club).role
-
-        self.assertTrue(role == 'Member')
-
-    def test_getCurrentUserClubs(self):
-        pass
