+++ backend/pytest.ini
@@ -1,0 +1,3 @@
+[pytest]
+DJANGO_SETTINGS_MODULE = ultimate_frisbee_management.settings
+python_files = tests.py test_*.py *_tests.py
